(WIP) refactor frontend
