+++ packages/app/src/components/catalog/EntityPage/Content/Docs.tsx
@@ -1,0 +1,13 @@
+import { EntityTechdocsContent } from '@backstage/plugin-techdocs';
+import React from 'react';
+
+import { ReportIssue } from '@backstage/plugin-techdocs-module-addons-contrib';
+import { TechDocsAddons } from '@backstage/plugin-techdocs-react';
+
+export const TechdocsContent = () => (
+  <EntityTechdocsContent>
+    <TechDocsAddons>
+      <ReportIssue />
+    </TechDocsAddons>
+  </EntityTechdocsContent>
+);
