+++ packages/app/src/components/catalog/EntityPage/Content/index.tsx
@@ -1,0 +1,7 @@
+export * from './Cicd';
+export * from './Dependencies';
+export * from './Docs';
+export * from './GithubIssues';
+export * from './GithubPullRequests';
+export * from './Overview';
+export * from './SecurityInsights';
