+++ packages/app/src/components/catalog/EntityPage/Pages/index.tsx
@@ -1,0 +1,8 @@
+export * from './Api';
+export * from './Component';
+export * from './DefaultEntity';
+export * from './Domain';
+export * from './Group';
+export * from './Resource';
+export * from './System';
+export * from './User';
